requirements.txt
# Core ML - TensorFlow approach (avoids PyTorch DLL issues)
tensorflow>=2.13.0
opencv-python>=4.8.0
numpy>=1.22,<=1.24.3
pillow>=10.0.0
streamlit>=1.36.0
pyyaml>=6.0.0
typing-extensions>=3.6.6,<4.6.0

# Alternative: PyTorch (if you want to try again)
# torch>=2.0.0
# torchvision>=0.15.0
# ultralytics==8.3.28
